feat:udpate typst 0.12 and tauri 2.0
