+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "typster",
   "private": true,
-  "version": "0.4.0",
+  "version": "0.12.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
