+++ src-tauri/capabilities/default.json
@@ -8,12 +8,14 @@
     "core:default",
     "fs:allow-read-file",
     "fs:allow-write-file",
+    "fs:allow-write-text-file",
     "fs:allow-read-dir",
     "fs:allow-copy-file",
     "fs:allow-mkdir",
     "fs:allow-remove",
     "fs:allow-rename",
     "fs:allow-exists",
+    "dialog:allow-save",
     {
       "identifier": "fs:scope",
       "allow": [
