+++ src-tauri/gen/schemas/capabilities.json
@@ -1,1 +1,1 @@
-{"main-capability":{"identifier":"main-capability","description":"Capability for the main window","local":true,"windows":["main"],"permissions":["core:default","fs:allow-read-file","fs:allow-write-file","fs:allow-read-dir","fs:allow-copy-file","fs:allow-mkdir","fs:allow-remove","fs:allow-rename","fs:allow-exists",{"identifier":"fs:scope","allow":["$HOME/**/*"]},{"identifier":"fs:allow-read-dir","allow":["$HOME/**/*"]},{"identifier":"fs:allow-read-file","allow":["$HOME/**/*"]},{"identifier":"fs:allow-write-file","allow":["$HOME/**/*"]},"core:window:allow-center","core:window:allow-set-position","core:window:allow-set-size","core:window:allow-set-maximizable","core:window:allow-set-minimizable","fs:default"]}}
+{"main-capability":{"identifier":"main-capability","description":"Capability for the main window","local":true,"windows":["main"],"permissions":["core:default","fs:allow-read-file","fs:allow-write-file","fs:allow-write-text-file","fs:allow-read-dir","fs:allow-copy-file","fs:allow-mkdir","fs:allow-remove","fs:allow-rename","fs:allow-exists","dialog:allow-save",{"identifier":"fs:scope","allow":["$HOME/**/*"]},{"identifier":"fs:allow-read-dir","allow":["$HOME/**/*"]},{"identifier":"fs:allow-read-file","allow":["$HOME/**/*"]},{"identifier":"fs:allow-write-file","allow":["$HOME/**/*"]},"core:window:allow-center","core:window:allow-set-position","core:window:allow-set-size","core:window:allow-set-maximizable","core:window:allow-set-minimizable","fs:default"]}}
